% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{build_tree}
\alias{build_tree}
\title{Proposes a candidate for th}
\usage{
build_tree(
  th,
  r,
  log_u,
  v,
  j,
  epsilon,
  log_p_old,
  log_p_th,
  grad_th,
  warmup,
  M
)
}
\description{
Proposes a candidate for th
}
