% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{find_reasonable_epsilon}
\alias{find_reasonable_epsilon}
\title{Calculate reasonable initial value for epsilon}
\usage{
find_reasonable_epsilon(
  th,
  log_p_th,
  grad_th,
  M,
  max_iter = 100,
  verbose = FALSE
)
}
\arguments{
\item{th}{a vector of parameters of interests}

\item{log_p_th}{a function that takes a vector of parameters and returns a log likelihood}

\item{grad_th}{a function that takes a vector of parameters and returns the gradient}

\item{M}{covariance of the momentums}

\item{max_iter}{max number of search iterationns}
}
\value{
a value for epsilon
}
\description{
Calculate reasonable initial value for epsilon
}
