% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_iteration}
\alias{hmc_iteration}
\title{Run one Hamiltonian Monte Carlo iteration}
\usage{
hmc_iteration(th, log_p_th, grad_th, epsilon, L, M)
}
\arguments{
\item{th}{a vector of parameters of interests}

\item{log_p_th}{a function that takes a vector of parameters and returns a log likelihood}

\item{grad_th}{a function that takes a vector of parameters and returns the gradient}

\item{epsilon}{step size}

\item{L}{an integer, trajectory length}

\item{M}{covariance for the momentum}
}
\value{
th: new sample of the parameters of interest
     p_jump: current iteration's acceptance probability
}
\description{
Code based on Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A., &
Rubin, D.B. (2013). Bayesian Data Analysis (3rd ed.). Chapman and Hall/R
}
