% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_run}
\alias{hmc_run}
\title{Run many chains of a Hamiltonian Monte Carlo sampler}
\usage{
hmc_run(log_p_th, grad_th, starting_values, iter, epsilon_0, L_0, M)
}
\arguments{
\item{log_p_th}{a function that takes a vector of parameters and returns a log likelihood}

\item{grad_th}{a function that takes a vector of parameters and returns the gradient}

\item{starting_values}{c x d matrix of starting values, c: number of chains, d: number of parameters}

\item{iter}{number of MCMC iterations}

\item{epsilon_0}{initial step size}

\item{L_0}{an integer, initial trajectory length}

\item{M}{covariance for the momentum}
}
\value{
th: new sample of the parameters of interest
     p_jump: current iteration's acceptance probability
}
\description{
Run many chains of a Hamiltonian Monte Carlo sampler
}
