% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnuts.R
\docType{package}
\name{mcnuts}
\alias{mcnuts}
\title{mcnuts: A package with some gradient-based MCMC algorithms}
\description{
This package implements Hamiltonion Monte Carlo and the No-U-Turn Sampler
for easy integration into a custom MCMC algorithm that might, for example,
include Gibbs steps.
}
