% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{nuts_iteration}
\alias{nuts_iteration}
\title{Run one NUTS iteration}
\usage{
nuts_iteration(
  th,
  log_p_th,
  grad_th,
  epsilon,
  M,
  max_depth = 5,
  warmup = FALSE,
  verbose = FALSE,
  t = NULL,
  epsilon_bar = NULL,
  H = NULL,
  gamma = 0.05,
  t0 = 10,
  kappa = 0.75,
  delta = 0.65,
  mu_eps = NULL
)
}
\arguments{
\item{th}{a vector of parameters of interests}

\item{log_p_th}{a function that takes a vector of parameters and returns a log likelihood}

\item{grad_th}{a function that takes a vector of parameters and returns the gradient}

\item{epsilon}{step size}

\item{M}{covariance of the momentums}

\item{max_depth}{max depth for recursive leap-frog tree}

\item{warmup}{boolean of whether to update epsilon or not}
}
\value{
List of outputs
      th: a new sample of the vector of parmeters
      epsilon: step size for the next iteration
}
\description{
Run one NUTS iteration
}
