% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{nuts_run}
\alias{nuts_run}
\title{Run NUTS with many chains}
\usage{
nuts_run(
  log_p_th,
  grad_th,
  starting_values,
  iter,
  M = NULL,
  p_adapt = 0.5,
  epsilon_bar = 1,
  H = 0,
  gamma = 0.05,
  t0 = 10,
  kappa = 0.75,
  delta = 0.65,
  verbose = FALSE
)
}
\arguments{
\item{log_p_th}{a function that takes a vector of parameters and returns a log likelihood}

\item{grad_th}{a function that takes a vector of parameters and returns the gradient}

\item{starting_values}{c x d matrix of starting values, c: number of chains, d: number of parameters}

\item{iter}{number of MCMC iterations}

\item{M}{covariance for the momentum}

\item{p_adapt}{the fragtion of iterations where step size epsilon is adapted}

\item{epsilon_bar}{initial guess for epsilon}

\item{verbose}{boolean indicating to print progress or not}
}
\value{
sims: iter x c x d array of MCMC sample
     L: vector of trajectory length
     eps: vector of step size
}
\description{
Run NUTS with many chains
}
